#38: Deleted `getVisibleRowList()` completely and used Qt APIs.
tabKeyNavigtion is enabled by default for QTableView, The missing thing
was that after handling all special keypress, any other keypress showed
be passed to be handled by the QTableView's default handler.

So no need to have special handling for the arrow keys, it's handled
already by the QTableView.

Also this gives additional feature of quick search. The user can type
few key sequence in succession to jump to the next cell that matches
what the user typed.

getVisibleRowList() shall not calculate the list everytime it's called 
https://github.com/embedded-slam/siraj/issues/issue/38