#71: Feature complete.
The user can now select any single cell from the main view and click
`Ctrl+F` to open the filter view and populate it with the selected cell.

To unfilter a cell from the filter view, the user shall select the cell
and click `Ctrl+U` to unfilter it. The selected cell MUST be the same
cell that was used to add the cell to the filter view originally.


Task-Url: https://github.com/embedded-slam/siraj/issues/issues/issue/71