Add option to store include non-matching line in the table.
If the log contains some lines that doesn't match the table schema and
is mainly intended to add more info for the most recent line before it
that match the schema. This option allows adding the non matching line
to the table by building a fake schema based on the previous matching
line.